fix: Fix epoch handling
On Linux and MacOS performance.now() does not increase when the computer
sleeps. This change updates the epoch every time when the page becomes
active ensuring that the timestamp matches the current time even when
the computer and browser becomes active after a sleep period.

Details about the performance.now() bug:
https://developer.mozilla.org/en-US/docs/Web/API/Performance/now#ticking_during_sleep