feat: Do not submit standalone metadata events
When the user is inactive for a longer period of time only the same
metadata events (and header events) were submitted periodically.
But as long as the metadata stays the same there is no additional value
in submitting it multiple times when there is no other interaction.
This change modifies the behaviour so, that metadata events are
submitted periodically if the metadata event changes or if other events
are also captured.