feat: Do not submit standalone Metadata events
When the user is inactive for a longer period of time only the same
metadata events (and header events) were submitted periodically.
This change modifies the behaviour so, that if the user is inactive
for a longer period of time no data is submitted, and once the user
becomes active again a new metadata event is submitted.